Finished

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,19 @@
 import { useWorkoutContext } from "../hooks/useWorkoutsContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext} from '../hooks/useAuthContext';
 
 const WorkoutDetails = ({workout}) => {
   const {dispatch} = useWorkoutContext();
+  const {user} = useAuthContext();
 
   const handleDelete = async() =>{
+    if(!user){
+      return;
+    }
+
     const response =  await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers:{ 'Authorization': `Bearer ${user.token}`}
     })
     const json = await response.json();
     console.log(json);
